Refine calendar visuals and modularize components
